requirements.txt
Django
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pandas
requests
python-decouple>=3.8
